requirements.txt
amqp==2.3.2
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
billiard==3.5.0.3
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
coverage==4.5.1
decorator==4.3.0
Django==2.0.6
django-cors-headers==1.1.0
django-redis==4.9.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
eventlet==0.23.0
greenlet==0.4.13
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.0
kombu==4.2.1
more-itertools==4.2.0
mysqlclient==1.3.12
openpyxl==2.5.4
parso==0.2.1
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
py==1.5.3
Pygments==2.2.0
pytest==3.6.1
pytest-cov==2.5.1
pytest-django==3.3.0
pytest-ipdb==0.1
pytz==2018.4
redis==2.10.6
requests==2.19.1
selenium==3.3.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
vine==1.1.4
virtualenv==15.1.0
wcwidth==0.1.7
